wait what?

+++ ft_strcat.c
@@ -14,8 +14,8 @@
 
 char	*ft_strcat(char *s1, const char *s2)
 {
-	int	b;
-	int	c;
+	size_t	b;
+	size_t	c;
 
 	b = 0;
 	while (s1[b] != '\0')
